+++ bignum/src/commonMain/kotlin/com/ionspin/kotlin/bignum/integer/BigInteger.kt
@@ -63,6 +63,14 @@ class BigInteger internal constructor(wordArray: WordArray, requestedSign: Sign)
         return this
     }
 
+    fun getBackingArrayCopy(): WordArray {
+        return magnitude.copyOf()
+    }
+
+    fun getSign(): Sign {
+        return sign
+    }
+
     companion object : BigNumber.Creator<BigInteger>, BigNumber.Util<BigInteger>, ByteArrayDeserializable<BigInteger> {
         private val arithmetic: BigIntegerArithmetic = chosenArithmetic
 
@@ -72,6 +80,10 @@ class BigInteger internal constructor(wordArray: WordArray, requestedSign: Sign)
         override val TEN = BigInteger(arithmetic.TEN, Sign.POSITIVE)
 
         val LOG_10_OF_2 = log10(2.0)
+
+        fun createFromWordArray(wordArray: WordArray, requestedSign: Sign): BigInteger {
+            return BigInteger(wordArray, requestedSign)
+        }
 
         override fun parseString(string: String, base: Int): BigInteger {
             if (base < 2 || base > 36) {
@@ -460,6 +472,21 @@ class BigInteger internal constructor(wordArray: WordArray, requestedSign: Sign)
         return BigInteger(wordArray = this.magnitude, requestedSign = Sign.POSITIVE)
     }
 
+    fun factorial(): BigInteger {
+        var result = ONE
+        var element = ONE
+        val abs = this.abs()
+        while (element <= abs) {
+            result *= element
+            element = element.inc()
+        }
+        return if (this.isNegative) {
+            -result
+        } else {
+            result
+        }
+    }
+
     fun pow(exponent: BigInteger): BigInteger {
         if (exponent < ZERO)
             throw ArithmeticException("Negative exponent not supported with BigInteger")
@@ -606,7 +633,7 @@ class BigInteger internal constructor(wordArray: WordArray, requestedSign: Sign)
 
     override fun compareTo(other: Any): Int {
         if (other is Number) {
-            if (ComparisonWorkaround.isSpecialHandlingForFloatNeeded(other)) {
+            if (RuntimePlatform.currentPlatform() == Platform.JS) {
                 return javascriptNumberComparison(other)
             }
         }
@@ -631,9 +658,10 @@ class BigInteger internal constructor(wordArray: WordArray, requestedSign: Sign)
      * to check if it's a decimal or integer number before comparing.
      */
     private fun javascriptNumberComparison(number: Number): Int {
-        val float = number.toFloat()
+        val double = number.toDouble()
         return when {
-            float % 1 == 0f -> compare(fromLong(number.toLong()))
+            double > Long.MAX_VALUE -> { compare(parseString(double.toString())) } // This whole block can be removed after 1.6.20 and https://github.com/JetBrains/kotlin/pull/4364
+            double % 1 == 0.0 -> compare(fromLong(number.toLong()))
             else -> compareFloatAndBigInt(number.toFloat()) { compare(it) }
         }
     }
@@ -752,7 +780,7 @@ class BigInteger internal constructor(wordArray: WordArray, requestedSign: Sign)
             val firstBit = magnitude[1] shl 63
             (magnitude[0].toLong() or firstBit.toLong()) * signum()
         } else {
-            return magnitude[0].toLong()
+            return magnitude[0].toLong() * signum()
         }
     }
 
@@ -812,7 +840,12 @@ class BigInteger internal constructor(wordArray: WordArray, requestedSign: Sign)
 
     override fun doubleValue(exactRequired: Boolean): Double {
         if (exactRequired && this.abs() > Double.MAX_VALUE) {
-            throw ArithmeticException("Cannot convert to float and provide exact value")
+            println(this.abs())
+            println(Double.MAX_VALUE)
+            if (this.abs() > Double.MAX_VALUE) {
+                println("huh")
+            }
+            throw ArithmeticException("Cannot convert to double and provide exact value")
         }
         return this.toString().toDouble()
     }